Creates files for each Movie Content Handler
