+++ UpcomingMovies/Scenes/Movies/MovieList/ContentHandlers/MoviesContentHandlerProtocol.swift
@@ -1,0 +1,18 @@
+//
+//  MoviesContentHandlerProtocol.swift
+//  UpcomingMovies
+//
+//  Created by Alonso on 6/13/20.
+//  Copyright © 2020 Alonso. All rights reserved.
+//
+
+import Foundation
+import UpcomingMoviesDomain
+
+protocol MoviesContentHandlerProtocol {
+    
+    var displayTitle: String { get }
+    
+    func getMovies(page: Int, completion: @escaping (Result<[UpcomingMoviesDomain.Movie], Error>) -> Void)
+    
+}
